Add mockchannel plugin
